+++ selfdrive/selfdrived/selfdrived.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 import os
-import math
 import time
 import threading
 
@@ -8,8 +7,6 @@ import cereal.messaging as messaging
 
 from cereal import car, log
 from msgq.visionipc import VisionIpcClient, VisionStreamType
-from opendbc.car import ACCELERATION_DUE_TO_GRAVITY
-from opendbc.car.interfaces import ACCEL_MIN, ACCEL_MAX, ISO_LATERAL_ACCEL
 
 
 from openpilot.common.params import Params
@@ -19,6 +16,7 @@ from openpilot.common.gps import get_gps_location_service
 
 from openpilot.selfdrive.car.car_specific import CarSpecificEvents
 from openpilot.selfdrive.locationd.helpers import PoseCalibrator, Pose
+from openpilot.selfdrive.selfdrived.common import check_excessive_actuation
 from openpilot.selfdrive.selfdrived.events import Events, ET
 from openpilot.selfdrive.selfdrived.state import StateMachine
 from openpilot.selfdrive.selfdrived.alertmanager import AlertManager, set_offroad_alert
@@ -31,7 +29,6 @@ SIMULATION = "SIMULATION" in os.environ
 TESTING_CLOSET = "TESTING_CLOSET" in os.environ
 
 LONGITUDINAL_PERSONALITY_MAP = {v: k for k, v in log.LongitudinalPersonality.schema.enumerants.items()}
-MIN_EXCESSIVE_ACTUATION_COUNT = int(0.25 / DT_CTRL)
 
 ThermalStatus = log.DeviceState.ThermalStatus
 State = log.SelfdriveState.OpenpilotState
@@ -43,36 +40,6 @@ ButtonType = car.CarState.ButtonEvent.Type
 SafetyModel = car.CarParams.SafetyModel
 
 IGNORED_SAFETY_MODES = (SafetyModel.silent, SafetyModel.noOutput)
-
-
-def check_excessive_actuation(sm: messaging.SubMaster, CS: car.CarState, calibrated_pose: Pose, counter: int) -> tuple[int, bool, float]:
-  # CS.aEgo can be noisy to bumps in the road, transitioning from standstill, losing traction, etc.
-  # longitudinal
-  accel_calibrated = calibrated_pose.acceleration.x
-  excessive_long_actuation = sm['carControl'].longActive and (accel_calibrated > ACCEL_MAX * 2 or accel_calibrated < ACCEL_MIN * 2)
-
-  # lateral
-  yaw_rate = calibrated_pose.angular_velocity.yaw
-  roll = sm['liveParameters'].roll
-  roll_compensated_lateral_accel = (CS.vEgo * yaw_rate) - (math.sin(roll) * ACCELERATION_DUE_TO_GRAVITY)
-
-  excessive_lat_actuation = False
-  # print('vEgo', CS.vEgo, yaw_rate, roll)
-  # print('roll_compensated_lateral_accel', roll_compensated_lateral_accel)
-  if sm['carControl'].latActive:
-    if not CS.steeringPressed:
-      if abs(roll_compensated_lateral_accel) > ISO_LATERAL_ACCEL * 2:
-        excessive_lat_actuation = True
-
-  # livePose acceleration can be noisy due to bad mounting or aliased livePose measurements
-  livepose_valid = abs(CS.aEgo - accel_calibrated) < 2
-  # print('excessive_long_actuation', excessive_long_actuation, 'excessive_lat_actuation', excessive_lat_actuation, 'livepose_valid', livepose_valid)
-  counter = counter + 1 if livepose_valid and (excessive_long_actuation or excessive_lat_actuation) else 0
-
-  # if counter > 0:
-  #   print('counter', counter, excessive_long_actuation, excessive_lat_actuation, livepose_valid)
-
-  return counter, counter > MIN_EXCESSIVE_ACTUATION_COUNT, roll_compensated_lateral_accel
 
 
 class SelfdriveD:
